更新至20210516左右,修改样式和部分逻辑为第一次更新代码之后修改过的样式和逻辑 本地运行,还没上线运行

+++ vue_mogu_admin/src/components/TodoList/index.vue
@@ -128,12 +128,20 @@ export default {
 
     },
     deleteTodo(todo) {
-      var params = {};
-      params.uid = todo.uid;
-      deleteTodo(params).then(response => {
-        if (response.code == this.$ECode.SUCCESS) {
-          this.getTodoList();
-        }
+      this.$confirm("此操作将删除该事项, 是否继续?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      }).then(() => {
+        var params = {};
+        params.uid = todo.uid;
+        deleteTodo(params).then(response => {
+          if (response.code == this.$ECode.SUCCESS) {
+            this.getTodoList();
+          }
+        });
+      }).catch(() => {
+        that.$commonUtil.message.info("已取消删除")
       });
     },
     editTodo({ todo, value }) {
